requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
boto3==1.18.65
botocore==1.21.65
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
idna==3.2
jmespath==0.10.0
multidict==5.1.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
slack-bolt==1.9.1
slack-sdk==3.10.1
slackclient==2.9.3
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
